+++ src/App/auth/Login.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { Link, Redirect } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import LinearProgress from '@material-ui/core/LinearProgress';
-import { getLoggedInStatus, getLoginProgress, loginUser } from '../../store/slicers/user_login';
-import { getSessionInfo, loadSession } from '../../store/slicers/user_session';
+import { getLoggedInStatus, getLoginProgress, loginUser } from '../../store/slicers/userLogin';
+import { getSessionInfo, loadSession } from '../../store/slicers/userSession';
 import AppBar from '../components/AppBar';
 import './style/auth.css';
