+++ src/App/components/TrackDetails.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 import AppBar from './AppBar';
 import { getProducts, loadProducts } from '../../store/slicers/product';
-import { getSessionInfo, loadSession } from '../../store/slicers/user_session';
+import { getSessionInfo, loadSession } from '../../store/slicers/userSession';
 import './style/TrackDetails.css';
 import Stats from './Stats';
